+++ CHANGELOG.md
@@ -5,11 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [0.3.0] - 2026-02-10
 
 ### Added
 
-- `limit`/`offset` pagination for listing sessions and API keys by resource. [#4](https://github.com/hpopp/auth-manager/pull/4)
+- `limit`/`offset` pagination for listing sessions and API keys by subject. [#4](https://github.com/hpopp/auth-manager/pull/4)
+- Snapshot replication for stale followers. [#5](https://github.com/hpopp/auth-manager/pull/5)
+
+### Removed
+
+- Cleaned up various unused legacy functions. [#5](https://github.com/hpopp/auth-manager/pull/5)
 
 ## [0.2.0] - 2026-02-09